chore(warnings): knock out 81 of 82 deprecation warnings
- ScheduleApp navigation: split NavHost into typed-vs-string overloads (fixes runtime "Cannot find startDestination kotlin.String" crash on iOS).
- kotlinx-datetime 0.7.x rename sweep: dayOfMonth → day, monthNumber → month.ordinal (or +1 where 1-based was needed).
- Compose Material3: Icons.Default.{KeyboardArrowLeft,KeyboardArrowRight,Logout} → Icons.AutoMirrored.Filled.* with the corresponding imports.
- gitlive Firebase: every .update("k" to v, ...) call switched to .updateFields { "k" to v } (deprecated → replacement).
- Compose Multiplatform Preview: import org.jetbrains.compose.ui.tooling.preview.Preview in all commonMain preview files; @file:Suppress("DEPRECATION") on the 6 affected files since the deprecation message keeps firing for both AndroidX and JetBrains FQNs in this Compose 1.10/AGP 9 combo.
- Excluded androidx.compose.ui:ui-tooling-preview transitive dep from the AndroidX ui-tooling artifact so commonMain Preview resolves to the Compose Multiplatform copy on Android.
- Cleaned several "unnecessary safe call" hits: EventMapper user/appliance access, ApplianceDetails createdUser, FirebaseUsersRepositoryImpl currentUser logger.
- Dropped redundant `.toLong()` on `LocalDate.toEpochDays()` (already returns Long in 0.7.x).
- Suppressed: SwipeToDismissBoxState confirmValueChange (no clean replacement without UX rework); GoogleSignIn Android SDK (entire SDK deprecated in favor of Credential Manager — large rewrite); WindowCompat statusBarColor/navigationBarColor (replaced by edge-to-edge in code, but still set for backwards compat).
- Set `binaryOption("bundleId", "ru.dvfu.appliances.composeApp")` on the iOS framework targets so kotlinc stops complaining about implicit bundleId.
- -Xexpect-actual-classes compiler flag added at kotlin{} root, kills the 5 expect/actual beta warnings.
- Final tally: 1 warning remaining (KMPNotifier internal interop klib export — third-party).

+++ composeApp/src/commonMain/kotlin/ru/dvfu/appliances/compose/calendars/WeekCalendar.kt
@@ -71,7 +71,7 @@ fun EventCalendar(
         floatingActionButton = {
             if (!currentUser.isAnonymousOrGuest) {
                 FloatingActionButton(
-                    onClick = { navController.navigate(AddEventRoute(dateEpochDay = Clock.System.todayIn(TimeZone.currentSystemDefault()).toEpochDays().toLong())) })
+                    onClick = { navController.navigate(AddEventRoute(dateEpochDay = Clock.System.todayIn(TimeZone.currentSystemDefault()).toEpochDays())) })
                 { Icon(Icons.Default.Add, "") }
             }
         },
